refactor integrations hierarchy
