+++ src/integrations/__init__.py
@@ -1,4 +1,5 @@
 from .base import BaseIntegration, IntegrationError
+from .project_management import ProjectManagementIntegration
 from .motion import MotionIntegration
 
-__all__ = ["BaseIntegration", "IntegrationError", "MotionIntegration"]
+__all__ = ["BaseIntegration", "ProjectManagementIntegration", "IntegrationError", "MotionIntegration"]
